# docs/a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'Wavy Web API'
  description: 'Web Resources Specification (A7) for Wavy'
servers:
- url: http://db.fe.up.pt
  description: Production server
...
paths:
  /login:
    get:
      operationId: R101
      summary: 'R101: Login Page'
      description: 'Provide the login page. Access: NAUTH'
      tags:
        - 'M01: Authentication'
      responses:
        '200':
          description: 'Show the Login Page.'
    
    post:
      operationId: R102
      summary: 'R102: Login Action'
      description: 'Process the login information submitted. Access: NAUTH'
      tags:
        - 'M01: Authentication'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '302':
          description: 'Redirect after processing the login'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Login was successful.'
                  value: '/home'
                302Error:
                  description: 'The provided credentials do not match our records.'
                  value: '/login'
...
  /home:
    get:
      operationId: R110
      summary: 'R110: Access the Home Page'
      description: 'Provide access to the home page and timeline. Access: NAUTH, AUTH, ADMIN, OWN'
      tags:
        - 'M02: Users'
      responses: 
        '200':
          description: 'Successfully retrieved home page.'
        '401':
          description: 'Unauthorized access.'
...